Remaking some llines in order to fix bugs related to CrossEntropyLoss error function
Modigfications submited
	modified:   experiment_darts_wine.py
	modified:   winedataset.py
	modified:   ../model_search.py
	modified:   ../utils.py

+++ cnn/darts_for_wine/winedataset.py
@@ -33,7 +33,7 @@ class WinesDataset(Dataset):
                 self.labels.append(l)
 
             self.data = torch.FloatTensor(self.data)
-            self.labels = torch.IntTensor(self.labels)
+            self.labels = torch.LongTensor(self.labels)
 
         else:
             self.data = torch.from_numpy(data.reshape(data.shape[0], 1, data.shape[1], data.shape[2]))
